Change Product fields

+++ products/models.py
@@ -7,8 +7,8 @@ class Product(models.Model):
     title = models.CharField(max_length=200, verbose_name='Название')
     description = models.TextField(verbose_name='Описание')
     scope_of_application = models.CharField(max_length=200, verbose_name='Сфера применения')
-    diametr = models.FloatField(verbose_name='Диаметр')
-    length = models.FloatField(verbose_name='Длинна')
+    diametr = models.IntegerField(verbose_name='Диаметр')
+    length = models.IntegerField(verbose_name='Длинна')
     image = models.ImageField(null=True, default='', verbose_name='Изображение')
     card = models.ForeignKey(
         Card,
